+++ backend/app/services/AuthService.ts
@@ -1,7 +1,7 @@
 import UserRepository from "../repositories/UserRepository";
 import { type IRegisterBody } from "../types/bodies";
 import { MainClient } from "../database/clients";
-import AuthSchema from "../validations/AuthSchema";
+import AuthSchema from "./validations/AuthSchema";
 import Validator from "../utils/validator";
 import Encrypt from "../utils/encrypt";
